+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 9: Error Handling & Regular Expressions
     - Error Handling with Try...Catch
+    - Regular Expressions [Part 1] - Evaluation Functions
 - An example with error catching was added.
+- An example with evaluation functions was added.
 
 ## September 21, 2020
 - Modern JavaScript From The Beginning (Udemy)